Combine similar tests, remove duplicated code.
